--- src/components/HeaderContent/types.ts
@@ -1,3 +1,0 @@
-export type THeaderContent = {
-  phoneNumber: string;
-};
