+++ src/components/MobileMenu/types.ts
@@ -1,0 +1,3 @@
+export type TMobileMenu = {
+  phoneNumber: string;
+};
