--- src/components/ui/MobileMenu/types.tsx
@@ -1,5 +1,0 @@
-export type MobileMenuProps = {
-  phoneNumber: string;
-  handleClick?: () => void;
-  className?: string;
-};
